Create CodingContestTorontoParkingTickets2013.py
This is my winning entry in a coding contest. The contest objective was to write a program to figure out which Toronto street gave the city the most revenue from parking tickets in the first three days of January 2012. Entries were judged based upon accuracy, time taken to execute, and neatness of code.
(http://www.programmr.com/Python/TorontoTickets2013)